+++ eggs/AndroidNext/src/main/java/com/android_next/egg/AndroidNextTimelineDialog.kt
@@ -5,7 +5,6 @@ import android.content.res.Configuration
 import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
 import androidx.compose.foundation.horizontalScroll
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -16,9 +15,7 @@ import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
@@ -28,9 +25,14 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.drawWithCache
+import androidx.compose.ui.geometry.CornerRadius
+import androidx.compose.ui.geometry.Offset
+import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.graphics.ColorMatrix
@@ -38,14 +40,18 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalLayoutDirection
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.drawText
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.rememberTextMeasurer
 import androidx.compose.ui.unit.LayoutDirection
 import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
 import com.dede.android_eggs.util.CustomTabsBrowser
 import com.dede.basic.requireDrawable
 import com.google.accompanist.drawablepainter.rememberDrawablePainter
 import kotlinx.coroutines.launch
 import java.util.Calendar
-import kotlin.math.max
 import kotlin.math.min
 
 internal var androidNextDialogVisible by mutableStateOf(false)
@@ -82,7 +88,7 @@ fun AndroidNextTimelineDialog(
         text = {
             Column {
                 Text(
-                    text = AndroidNextEasterEgg.getTimelineMessage(context),
+                    text = getTimelineMessage(context),
                     style = MaterialTheme.typography.bodyMedium
                 )
                 Spacer(modifier = Modifier.height(12.dp))
@@ -109,40 +115,45 @@ fun AndroidNextTimelineDialog(
     )
 }
 
+private val offsetXPercentArr = floatArrayOf(
+    49 / 789f,
+    153 / 789f,
+    257 / 789f,
+    361 / 789f,
+    465 / 789f,
+    569 / 789f,
+    717 / 789f
+)
+
+private const val offsetYPercent = 135 / 180f
+
 @Composable
 private fun AndroidReleaseTimeline() {
-    val calendar = Calendar.getInstance()
-    val year = calendar.get(Calendar.YEAR)
-    val month = calendar.get(Calendar.MONTH)// [0, 11]
-    // Month    Progress    Calender.MONTH
-    // Feb          0           1
-    // ...
-    // Jul          5           6
-    // Aug          -           7
-    val offsetXArr = intArrayOf(20, 111, 202, 294, 386, 478, 584)
-    val nextReleaseYear = AndroidNextEasterEgg.RELEASE_YEAR
-    val offsetXIndex =
-        if (year < nextReleaseYear || (year == nextReleaseYear && month < Calendar.FEBRUARY)) {
+    val nowDate = Calendar.getInstance().setDateZero()
+    val releaseDate = remember { getReleaseCalendar() }
+
+    val offsetXIndex = remember(nowDate, releaseDate) {
+        val diffMonth = getDateDiffMonth(start = nowDate, end = releaseDate)
+        if (diffMonth > MONTH_CYCLE) {
             // No preview
             -1
-        } else if (year == nextReleaseYear && month in Calendar.FEBRUARY..Calendar.JULY) {
+        } else if (diffMonth < MONTH_CYCLE) {
             // Preview
-            month - 1
+            MONTH_CYCLE - diffMonth - 1
         } else {
             // Final release
             6
         }
+    }
     val isFinalRelease = offsetXIndex == 6
-    val hasPreview = offsetXIndex != -1
-    val offsetX = offsetXArr[min(offsetXArr.size - 1, max(offsetXIndex, 0))]
 
     val scrollState = rememberScrollState()
-    if (hasPreview) {
+    if (offsetXIndex != -1) {
         LaunchedEffect(offsetXIndex, isFinalRelease) {
             val value = if (isFinalRelease) {
                 scrollState.maxValue
             } else {
-                scrollState.maxValue / (offsetXArr.size) * (offsetXIndex)
+                scrollState.maxValue / offsetXPercentArr.size * offsetXIndex
             }
             launch {
                 scrollState.animateScrollTo(value)
@@ -156,45 +167,83 @@ private fun AndroidReleaseTimeline() {
                 .height(160.dp)
                 .aspectRatio(789f / 180)
         ) {
-            if (hasPreview) {
-                Box(
-                    modifier = Modifier
-                        .padding(top = 103.dp, start = offsetX.dp)
-                ) {
-                    val shape = RoundedCornerShape(
-                        topStartPercent = 50, topEndPercent = 50,
-                        bottomEndPercent = 50, bottomStartPercent = 50
-                    )
-                    if (isFinalRelease) {
-                        // Final release
-                        Box(
-                            modifier = Modifier
-                                .width(106.dp)
-                                .height(34.dp)
-                                .background(Color(0xFF3DDC84), shape)
-                        )
-                    } else {
-                        // Preview
-                        Box(
-                            modifier = Modifier
-                                .width(52.dp)
-                                .height(34.dp)
-                                .background(Color(0xFFF86734), shape)
-                        )
-                    }
-                }
-            }
+            val context = LocalContext.current
             val matrix = ColorMatrix()
-            if (isSystemNightMode(LocalContext.current)) {
+            if (isSystemNightMode(context)) {
                 // Increase the overall brightness and more blue brightness
                 matrix.setToScale(1.3f, 1.5f, 2f, 1f)
             }
+
+            val timelineMonths = remember { getReleaseCycleMonths(context) }
+            val textMeasurer = rememberTextMeasurer(cacheSize = timelineMonths.size)
             Image(
                 painter = painterResource(id = R.drawable.timeline_bg),
                 contentDescription = null,
-                colorFilter = ColorFilter.colorMatrix(matrix)
+                colorFilter = ColorFilter.colorMatrix(matrix),
+                modifier = Modifier.drawWithCache {
+                    onDrawWithContent {
+                        for ((index, month) in timelineMonths.withIndex()) {
+                            val isLastMonth = index == timelineMonths.size - 1
+
+                            val textLayout = textMeasurer.measure(
+                                text = month,
+                                style = TextStyle(
+                                    fontSize = 14.sp,
+                                    fontWeight = if (isLastMonth) FontWeight.Bold else FontWeight.Medium
+                                ),
+                            )
+
+                            val offsetX = size.width * offsetXPercentArr[index]
+                            val offsetY = size.height * offsetYPercent
+
+                            if (index == offsetXIndex) {
+                                val rectSize = Size(
+                                    width = textLayout.size.width + textLayout.size.height * 1.3f,
+                                    height = textLayout.size.height * 1.6f
+                                )
+                                val radius = min(rectSize.height, rectSize.width) / 2f
+                                drawRoundRect(
+                                    color = if (isLastMonth)
+                                        Color(0xFF3DDC84)
+                                    else
+                                        Color(0xFFF86734),
+                                    topLeft = Offset(
+                                        x = offsetX - rectSize.width / 2f,
+                                        y = offsetY - rectSize.height / 2f
+                                    ),
+                                    size = rectSize,
+                                    cornerRadius = CornerRadius(radius, radius)
+                                )
+                            }
+
+                            drawText(
+                                textLayoutResult = textLayout,
+                                color = if (isLastMonth)
+                                    Color(0xFF188038)
+                                else
+                                    Color.Black,
+                                topLeft = Offset(
+                                    x = offsetX - textLayout.size.width / 2f,
+                                    y = offsetY - textLayout.size.height / 2f,
+                                ),
+                            )
+                        }
+
+                        drawContent()
+                    }
+                }
             )
         }
+    }
+}
+
+private fun getTimelineMessage(context: Context): String {
+    val nowDate = Calendar.getInstance().setDateZero()
+    val releaseDate = getReleaseCalendar()
+    return if (nowDate.after(releaseDate)) {
+        context.getString(R.string.summary_android_release_pushed)
+    } else {
+        context.getString(R.string.summary_android_waiting)
     }
 }